backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.0
pydantic==2.9.0
python-dotenv==1.0.1
firebase-admin==6.5.0
pytesseract==0.3.13
Pillow==11.1.0
groq==0.25.0
httpx>=0.27.0
python-multipart==0.0.12
